[NZT-130] Refactor Roll
